requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.7.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
aiosqlite==0.20.0
